Adding color indicator for log displayed in the console.

+++ Hammann/source/Config.hpp
@@ -29,8 +29,8 @@ static const std::string ResourcePathRoot = "resource";
 
 static const bool LogResourceFileDetail = true;
 
-static const float scaleMax = 0.5;
-static const float scaleMin = 0.2;
+static const double scaleMax = 0.5;
+static const double scaleMin = 0.2;
 
 static const std::string LogfilePath = "log\\log.txt";
